Move MultipleCheckbox.jsx to GraphView
